+++ src/com/songlib/Controller.java
@@ -1,4 +1,4 @@
-package sample;
+package com.songlib;
 
 public class Controller {
 }
